Add query folder and group all query object types
