--- app/graphql/types/answer_type.rb
@@ -1,6 +1,0 @@
-module Types
-  class AnswerType < Types::BaseObject
-    field :id, ID, null: false
-    field :value, String, null: false
-  end
-end
