+++ app/graphql/types/query/answer_type.rb
@@ -1,0 +1,8 @@
+module Types
+  module Query
+    class AnswerType < Types::BaseObject
+      field :id, ID, null: false
+      field :value, String, null: false
+    end
+  end
+end
